ISPN-78 - Large Object Support
Instead of only the key store for each Chunk in LargeObjectMetadata a
newly to be implemented ChunkMetadata instance. For the time being it
will only hold a Chunk's key and its size. In the future more data
might be added.

Replace ChunkKeyProducer with KeyGenerator as it does not offer any
advantages over the latter.

Do not prefix chunk keys with a configurable string. In the future
there might be KeyGenerators that do not produce string chunk keys.